refactor(utils): move adding plugin to addEntries
