+++ lib/utils/addEntries.js
@@ -2,6 +2,7 @@
 
 const webpack = require('webpack');
 const createDomain = require('./createDomain');
+const getSocketClientPath = require('./getSocketClientPath');
 const DevServerEntryPlugin = require('./DevServerEntryPlugin');
 
 /**
@@ -115,11 +116,23 @@ function addEntries(compiler, options, server) {
     return defaultValue;
   };
 
+  const findHMRPlugin = (config) => {
+    if (!config.plugins) {
+      // eslint-disable-next-line no-undefined
+      return undefined;
+    }
+
+    return config.plugins.find(
+      (plugin) => plugin.constructor === webpack.HotModuleReplacementPlugin
+    );
+  };
+
   const compilers = compiler.compilers || [compiler];
 
   // eslint-disable-next-line no-shadow
   compilers.forEach((compiler) => {
     const config = compiler.options;
+    config.plugins = config.plugins || [];
     /** @type {Boolean} */
     const webTarget = [
       'web',
@@ -167,16 +180,19 @@ function addEntries(compiler, options, server) {
       compiler.hooks.entryOption.call(config.context, config.entry);
     }
 
-    if (options.hot || options.hot === 'only') {
-      config.plugins = config.plugins || [];
-      if (
-        !config.plugins.find(
-          // Check for the name rather than the constructor reference in case
-          // there are multiple copies of webpack installed
-          (plugin) => plugin.constructor.name === 'HotModuleReplacementPlugin'
-        )
-      ) {
-        config.plugins.push(new webpack.HotModuleReplacementPlugin());
+    const providePlugin = new webpack.ProvidePlugin({
+      __webpack_dev_server_client__: getSocketClientPath(options),
+    });
+    providePlugin.apply(compiler);
+
+    if (!findHMRPlugin(compiler.options)) {
+      // add and apply the HMR plugin, if it didn't exist before.
+      if (options.hot === true || options.hot === 'only') {
+        config.plugins = config.plugins || [];
+
+        const plugin = new webpack.HotModuleReplacementPlugin();
+        config.plugins.push(plugin);
+        plugin.apply(compiler);
       }
     }
   });
